Update front end when update reactjs and backend structure

+++ src/Server/Application/Camino.Application.Contracts/AppServices/Media/Dtos/UserPhotoResult.cs
@@ -9,7 +9,7 @@
         public string Description { get; set; }
         public string ImageData { get; set; }
         public long UserId { get; set; }
-        public byte TypeId { get; set; }
+        public int TypeId { get; set; }
         public byte[] BinaryData { get; set; }
     }
 }
