+++ src/Server/Shared/Camino.Shared/Enums/UserPictureTypes.cs
@@ -1,6 +1,6 @@
 ﻿namespace Camino.Shared.Enums
 {
-    public enum UserPictureTypes : byte
+    public enum UserPictureTypes : int
     {
         Undefined = 0,
         Avatar = 1,
